start of util.c splitup, created coordinateutils
