+++ common/util.h
@@ -698,7 +698,6 @@ int UtilSetSlaveObject(ObjectPosition *OP, char *Filename, char debug);
 int UtilSetAdaptiveSyncPoint(AdaptiveSyncPoint *ASP, FILE *filefd, char debug);
 void UtilSetObjectPositionIP(ObjectPosition *OP, char *IP);
 
-void llhOffsetMeters(double *llh, const double *xyzOffset);
 void llhToXyz(double lat, double lon, double height, double *x, double *y, double *z);
 void enuToLlh(const double *iLlh, const double *xyz, double *llh);
 void createEnuMatrix(double lat, double lon, double *enuMat);
