+++ common/util/coordinateutils.hpp
@@ -1,0 +1,1 @@
+void llhOffsetMeters(double *llh, const double *xyzOffset);
